requirements.txt
beautifulsoup4==4.9.3
fake_useragent==1.1.0
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dotenv==1.0.0
python_dateutil==2.8.2
Requests==2.31.0
SQLAlchemy==1.4.40
urllib3==1.26.12
